pyproject.toml
[tool.poetry]
name = "repl-nix-subtitlesnaps"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.50.2"
streamlit = "^1.38.0"
pydub = "^0.25.1"
moviepy = "^1.0.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
